refactor: move inline back into the fold
