+++ demo/bar.html
@@ -52,7 +52,7 @@
     </head>
     <body>
 
-        <rh-switcher key="demo" on-message="Viewing the new design" off-message="Viewing the original design">
+        <rh-switcher variant="bar" key="demo" on-message="Viewing the new design" off-message="Viewing the original design">
             <pfe-icon icon="circle-info" size="sm" aria-label="Awards"></pfe-icon> This content is in open beta
         </rh-switcher>
         
@@ -60,7 +60,7 @@
             <a href="/" slot="logo">Open Beta</a>
             <ul slot="nav">
                 <li><a href="/demo/">Card</a></li>
-                <li><a href="/demo/topbar.html">Top Bar</a></li>
+                <li><a href="/demo/bar.html">Bar</a></li>
                 <li><a href="/demo/inline.html">Inline</a></li>
             </ul>
         </rh-secondary-nav>
