+++ demo/index.html
@@ -51,7 +51,7 @@
         <script src="demo.js" type="module"></script>
     </head>
     <body>
-        <rh-switcher key="demo" on-message="Viewing the new design" off-message="Viewing the original design" card>
+        <rh-switcher variant="card" key="demo" on-message="Viewing the new design" off-message="Viewing the original design" card>
             <h2>New Design Option</h2>
             <rh-feedback-form slot="form"></rh-feedback-form>
         </rh-switcher>
@@ -59,7 +59,7 @@
             <a href="/" slot="logo">Open Beta</a>
             <ul slot="nav">
                 <li><a href="/demo/">Card</a></li>
-                <li><a href="/demo/topbar.html">Top Bar</a></li>
+                <li><a href="/demo/bar.html">Bar</a></li>
                 <li><a href="/demo/inline.html">Inline</a></li>
             </ul>
         </rh-secondary-nav>
