+++ demo/lightdom.css
@@ -5,7 +5,7 @@ html, body {
     font-size: 1rem;
 }
 
-rh-switcher:not([card]):not(:defined) {
+rh-switcher[variant="bar"]:not(:defined) {
     display: grid;
     background: var(--rh-switcher-background-color, #bde1f4);
     min-height: 3.4rem;
